ci(Travis): prevent inconsequential build failures
Prevents Travis from failing the build by allowing Semantic Release pre script to throw without
failing the build. Because the build depends on the Semantic Release pre script for its version
number, we run two builds. The first is a development build to ensure the build functions as
expected, giving Travis an opportunity to bail. The latter is run in a later build step following
the pre script version update.